Migrate most of the remaining tests.
The only tests still suffixed `.old.c` are side-tracing tests, which I
haven't even looked at, since side-tracing is broken for now.
