--- tests/c/constexpr.c.O0
@@ -1,2 +1,0 @@
-// ## -O0 doesn't make a constant expression.
-// ignore-if: true
