+++ tests/c/inline_asm.c
@@ -1,7 +1,7 @@
 // Run-time:
 //   env-var: YKD_LOG_IR=-:jit-pre-opt
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   status: error
 
 // Check that we can handle inline asm properly (currently expectely fails
