--- tests/c/lifetime_intrinsics.c.O0
@@ -1,2 +1,0 @@
-// # no lifetime annotations at -O0.
-// ignore-if: true
