+++ tests/c/many_threads_one_loc.c
@@ -3,7 +3,7 @@
 // Run-time:
 //   env-var: YKD_LOG_IR=-:aot
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   stderr:
 //     ...
 //     --- Begin aot ---
@@ -15,7 +15,7 @@
 //     }
 //     ...
 //     --- End aot ---
-//     jitstate: enter-jit-code
+//     yk-jit-event: enter-jit-code
 //     ...
 
 // Check that compiling and running traces in parallel works.
