+++ tests/c/nested_writetoptr.c
@@ -1,19 +1,19 @@
 // Run-time:
 //   env-var: YKD_LOG_IR=-:aot
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   stderr:
-//     jitstate: start-tracing
-//     jitstate: stop-tracing
+//     yk-jit-event: start-tracing
+//     yk-jit-event: stop-tracing
 //     --- Begin aot ---
 //     ...
-//     define dso_local void @foo(...
+//     func foo(...
 //     ...
-//     store ptr %{{num1}}, ptr @shadowstack_0, align 8...
+//     *@shadowstack_0 = %{{1_1}}
 //     ...
-//     call void @bar(...
+//     call bar(...
 //     ...
-//     store ptr %{{num2}}, ptr @shadowstack_0, align 8...
+//     *@shadowstack_0 = %{{0_1}}
 //     ...
 //     --- End aot ---
 //     ...
