+++ tests/c/no_trace_annotation.c
@@ -1,56 +1,55 @@
 // Run-time:
-//   env-var: YKD_LOG_JITSTATE=-
-//   env-var: YKD_SERIALISE_COMPILATION=1
 //   env-var: YKD_LOG_IR=-:aot,jit-pre-opt,jit-post-opt
+//   env-var: YKD_SERIALISE_COMPILATION=1
+//   env-var: YK_LOG=255
 //   stderr:
 //     ...
 //     --- Begin aot ---
 //     ...
-//     store i32 2...
-//     ...
-//     ...add...
+//     call call_me()...
 //     ...
 //     --- End aot ---
+//     ...
 //     --- Begin jit-pre-opt ---
 //     ...
-//     store i32 2...
-//     ...
-//     ...add...
+//     call @call_me()...
 //     ...
 //     --- End jit-pre-opt ---
-//     --- Begin jit-post-opt ---
-//     ...
-//     store i32 2...
-//     ...
-//     ...add...
-//     ...
-//     --- End jit-post-opt ---
+//     Can't JIT this!
+//     yk-jit-event: enter-jit-code
+//     Can't JIT this!
+//     Can't JIT this!
+//     Can't JIT this!
+//     yk-jit-event: deoptimise
 //     ...
 
-// Check that our NOOPT_VAL macro blocks optimisations.
+// Check that the `yk_outline` annotation works.
 
 #include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <yk.h>
 #include <yk_testing.h>
+
+__attribute__((yk_outline)) void call_me(void) {
+  fprintf(stderr, "Can't JIT this!\n");
+}
 
 int main(int argc, char **argv) {
   YkMT *mt = yk_mt_new(NULL);
   yk_mt_hot_threshold_set(mt, 0);
   YkLocation loc = yk_location_new();
 
-  int x, res = 0, i = 3;
+  int i = 5;
+  NOOPT_VAL(loc);
   NOOPT_VAL(i);
   while (i > 0) {
     yk_mt_control_point(mt, &loc);
-    x = 2;
-    NOOPT_VAL(x);
-    res = x + 3; // We don't want this operation to be optimised away.
+    call_me(); // This call must not be inlined.
     i--;
   }
-  NOOPT_VAL(res);
-  assert(res == 5);
+
   yk_location_drop(loc);
   yk_mt_shutdown(mt);
   return (EXIT_SUCCESS);
