+++ tests/c/no_trace_annotation2.c
@@ -1,37 +1,30 @@
 // Run-time:
 //   env-var: YKD_LOG_IR=-:aot,jit-pre-opt,jit-post-opt
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   stderr:
 //     ...
 //     --- Begin aot ---
 //     ...
-//     call void @call_me()...
-//     ...
-//     ..."yk_outline"...
+//     call call_me()...
 //     ...
 //     --- End aot ---
 //     ...
 //     --- Begin jit-pre-opt ---
 //     ...
-//     call void @call_me()...
+//     call @call_me()...
 //     ...
 //     --- End jit-pre-opt ---
-//     ...
-//     --- Begin jit-post-opt ---
-//     ...
-//     tail call void @call_me()...
-//     ...
-//     --- End jit-post-opt ---
-//     ...
-//     jitstate: enter-jit-code
+//     Can't JIT this!
+//     Or this!
+//     yk-jit-event: enter-jit-code
 //     Can't JIT this!
 //     Or this!
 //     Can't JIT this!
 //     Or this!
 //     Can't JIT this!
 //     Or this!
-//     jitstate: deoptimise
+//     yk-jit-event: deoptimise
 //     ...
 
 // Check that the `yk_outline` annotation works when a `yk_outline` annotated
