+++ tests/c/not_loopy_funcs_inlined_by_default.c
@@ -5,18 +5,14 @@
 //     ...
 //     --- Begin jit-pre-opt ---
 //     ...
-//     define ptr @__yk_compiled_trace_0(ptr %0, ptr %1) {
-//       ...
-//       call void @never_inline_into_trace(i32 noundef ...
-//       ...
-//     }
+//     %{{12}}: i32 = call @call_me(%{{6}})
 //     ...
 //     --- End jit-pre-opt ---
 
-// Check that `yk_outline` wins over `yk_unroll_safe`.
+// Check that functions containing no loops get inlined into the trace.
 //
-// Although it is tempting to have clang emit an error when these conflicting attributes are used
-// together, the idiomatic "clang way" is to have one attribute win out over the other.
+// We can only see a call to `call_me()` in the trace if `inline_into_trace()`
+// itself was also inlined into the trace.
 
 #include <assert.h>
 #include <stdio.h>
@@ -28,12 +24,8 @@
 
 int call_me(int x); // from extra_linkage/call_me.c
 
-// Both `yk_outline` and `yk_unroll_safe`!
-__attribute__((yk_outline, yk_unroll_safe)) void
-never_inline_into_trace(int x) {
-  while (x--)
-    call_me(x);
-}
+// A function with no loops.
+__attribute__((noinline)) void inline_into_trace(int x) { call_me(x); }
 
 int main(int argc, char **argv) {
   YkMT *mt = yk_mt_new(NULL);
@@ -45,7 +37,7 @@ int main(int argc, char **argv) {
   NOOPT_VAL(i);
   while (i > 0) {
     yk_mt_control_point(mt, &loc);
-    never_inline_into_trace(i);
+    inline_into_trace(i);
     i--;
   }
   yk_location_drop(loc);
