--- tests/c/not_loopy_funcs_inlined_by_default.c.O0
@@ -1,14 +1,0 @@
-// Run-time:
-//   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_IR=-:jit-pre-opt
-//   stderr:
-//     ...
-//     --- Begin jit-pre-opt ---
-//     ...
-//     define ptr @__yk_compiled_trace_0(ptr %0, ptr %1) {
-//       ...
-//       %{{12}} = call i32 @call_me(i32...
-//       ...
-//     }
-//     ...
-//     --- End jit-pre-opt ---
