+++ tests/c/reentrant.c
@@ -5,31 +5,18 @@
 // ## mappable block :           <---
 // ##        unmappable block       |
 // ##         mappable block  ------
-// ignore-if: test "$YKB_TRACER" = "swt"
+// ##
+// ## Also the new trace builder cannot yet handle callbacks -- hits a todo!
+// ignore-if: test "$YKB_TRACER" = "swt" || true
 // Run-time:
 //   env-var: YKD_SERIALISE_COMPILATION=1
 //   env-var: YKD_LOG_IR=-:jit-pre-opt
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 
 // Check that we can reliably deal with "foreign" (not compiled with ykllvm)
 // code that calls back into "native code".
-
-// FIXME:  When the trace compiler encounters a call to foreign code, it
-// simply emits a `call` into the JIT trace (not inlining it -- how could it?
-// We have no IR for the foreign code). However, in order to continue
-// building the remainder of the JIT trace after the call, the trace compiler
-// must skip over all the parts of the of "input trace" (e.g. the hardware
-// trace) that correspond with the outlined call. Currently the compiler
-// simply waits until it finds a block with IR again. This is clearly
-// incorrect if foreign code calls back to code we have IR for.
 //
-// The good news is that there is an assertion checking that the place the
-// trace compiler tries to pick up compilation again is as expected. This
-// ensures that we don't miscompile at least. This test fails that assertion:
-//
-// Assertion `std::get<1>(ResumeAfter.getValue())->getParent() == BB' failed.
-//
-// (When fixing this test, add lines to match in the output)
+// FIXME: actually match some IR/output.
 
 #include <assert.h>
 #include <stdio.h>
