+++ tests/c/simple_non_serialised.c
@@ -1,17 +1,15 @@
 // Run-time:
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   env-var: YKD_LOG_STATS=/dev/null
 //   stderr:
-//     jitstate: start-tracing
+//     yk-jit-event: start-tracing
 //     i=4
-//     jitstate: stop-tracing
+//     yk-jit-event: stop-tracing
 //     i=3
-//     jitstate: enter-jit-code
+//     yk-jit-event: enter-jit-code
 //     i=2
 //     i=1
-//     jitstate: deoptimise
-//   stdout:
-//     exit
+//     yk-jit-event: deoptimise
 
 // Check that basic trace compilation in a thread works.
 
@@ -29,21 +27,16 @@ int main(int argc, char **argv) {
   yk_mt_hot_threshold_set(mt, 0);
   YkLocation loc = yk_location_new();
 
-  int res = 9998;
   int i = 4;
   NOOPT_VAL(loc);
-  NOOPT_VAL(res);
   NOOPT_VAL(i);
   while (i > 0) {
     yk_mt_control_point(mt, &loc);
     if (i == 3)
       __ykstats_wait_until(mt, test_compiled_event);
     fprintf(stderr, "i=%d\n", i);
-    res += 2;
     i--;
   }
-  printf("exit");
-  NOOPT_VAL(res);
   yk_location_drop(loc);
   yk_mt_shutdown(mt);
   return (EXIT_SUCCESS);
