+++ tests/c/smmultisrc2.c
@@ -1,7 +1,9 @@
+// ## not yet implemented: ran out of gp regs
+// ignore-if: true
 // Run-time:
 //   env-var: YKD_LOG_IR=-:aot
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   stdout:
 //     1 2 3 4 5 6
 //     1 2 3 4 5 6
