+++ tests/c/stats1.c
