+++ tests/c/stats2.c
