+++ tests/c/stats4.c
