+++ tests/c/struct_simple.c
@@ -1,19 +1,19 @@
 // Run-time:
 //   env-var: YKD_LOG_IR=-:jit-pre-opt
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   stderr:
 //     ...
-//     jitstate: start-tracing
+//     yk-jit-event: start-tracing
 //     3:1
-//     jitstate: stop-tracing
+//     yk-jit-event: stop-tracing
 //     --- Begin jit-pre-opt ---
 //     ...
 //     --- End jit-pre-opt ---
 //     2:1
-//     jitstate: enter-jit-code
+//     yk-jit-event: enter-jit-code
 //     1:1
-//     jitstate: deoptimise
+//     yk-jit-event: deoptimise
 //     ...
 
 // Check that we can handle struct field accesses.
