+++ tests/c/trace_too_long_hwt.c
@@ -1,14 +1,13 @@
-// ## This tests that traces that generate too many blocks cause "trace too
-// ## long" warnings. This can be very slow (e.g. on swt), so ignore it except
-// ## where we know it'll run fast enough.
-// ignore-if: test "$YKB_TRACER" != "hwt"
+// ## This test checks that an overflowing PT buffer is caught at the point
+// ## where a trace is stopped, not after trace mapping. It therefore only works
+// ## on hwt.
+// ignore-if: test "$YKB_TRACER" = "swt"
 // Run-time:
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YK_LOG=255
 //   stderr:
 //     ...
-//     jitstate: stop-tracing
-//     jitstate: trace-compilation-aborted: Trace too long
+//     yk-warning: stop-tracing-aborted: Trace too long
 
 #include <assert.h>
 #include <stdio.h>
@@ -23,18 +22,18 @@ int main(int argc, char **argv) {
   YkLocation loc1 = yk_location_new();
   YkLocation loc2 = yk_location_new();
 
-  int i = 2000;
+  int i = 100000;
   NOOPT_VAL(loc1);
   NOOPT_VAL(loc2);
   NOOPT_VAL(i);
   YkLocation *loc = &loc1;
   while (i > 0) {
     yk_mt_control_point(mt, loc);
-    if (i == 2000)
+    if (i == 100000)
       loc = &loc2;
     else if (i == 2)
       loc = &loc1;
-    fprintf(stdout, "i=%d\n", i);
+    NOOPT_VAL(i);
     i--;
   }
   printf("exit");
