+++ tests/c/unroll_safe_implies_noinline.c
@@ -1,25 +1,32 @@
-// ignore-if: test "$YKB_TRACER" != "hwt"
 // Run-time:
 //   env-var: YKD_SERIALISE_COMPILATION=1
-//   env-var: YKD_LOG_JITSTATE=-
+//   env-var: YKD_LOG_IR=-:aot
 //   stderr:
-//     jitstate: start-tracing
-//     set jump point
-//     jumped!
-//     jitstate: stop-tracing
-//     jitstate: trace-compilation-aborted: Encountered longjmp
+//     ...
+//     --- Begin aot ---
+//     ...
+//     call never_aot_inline(...
+//     ...
+//     --- End aot ---
 //     ...
 
-// Check that we bork on a call to setjmp in unmapped code.
+// Check that `yk_unroll_safe` implies `noinline` (thus blocking AOT inlining).
 
 #include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <unistd.h>
 #include <yk.h>
 #include <yk_testing.h>
 
-void unmapped_setjmp(void);
+int call_me(int x); // from extra_linkage/call_me.c
+
+// A function containing a loop and marked `yk_unroll_safe`.
+__attribute__((yk_unroll_safe)) void never_aot_inline(int x) {
+  while (x--)
+    call_me(x);
+}
 
 int main(int argc, char **argv) {
   YkMT *mt = yk_mt_new(NULL);
@@ -31,10 +38,9 @@ int main(int argc, char **argv) {
   NOOPT_VAL(i);
   while (i > 0) {
     yk_mt_control_point(mt, &loc);
-    unmapped_setjmp();
+    never_aot_inline(i);
     i--;
   }
-
   yk_location_drop(loc);
   yk_mt_shutdown(mt);
   return (EXIT_SUCCESS);
