+++ tests/c/unroll_safe_inlines.c
@@ -5,11 +5,7 @@
 //     ...
 //     --- Begin jit-pre-opt ---
 //     ...
-//     define ptr @__yk_compiled_trace_0(ptr %0, ptr %1, ptr %2) {
-//       ...
-//       %{{12}} = tail call i32 @call_me(i32...
-//       ...
-//     }
+//     %{{12}}: i32 = call @call_me(...
 //     ...
 //     --- End jit-pre-opt ---
